* FIX [warning] fix return type conflict.

+++ include/nng/supplemental/nanolib/file.h
@@ -9,7 +9,7 @@ extern bool    nano_file_exists(const char *fpath);
 extern char *  nano_getcwd(char *buf, size_t size);
 extern int64_t nano_getline(
     char **restrict line, size_t *restrict len, FILE *restrict fp);
-extern int file_write_string(const char *fpath, const char *string);
-extern int file_load_data(const char *filepath, void **data);
+extern int    file_write_string(const char *fpath, const char *string);
+extern size_t file_load_data(const char *filepath, void **data);
 
 #endif
